Refactor ScriptSetupGenerator to accept permutations in constructor
The ScriptSetupGenerator has been refactored to accept permutations as a parameter in the constructor, rather than in the generate method. This change has therefore been reflected in the ComponentSetProcessor, where the new ScriptSetupGenerator object now takes in permutations when instantiated and no longer holds an instance of ScriptSetupGenerator as a private property.