Implement token replacement functionality in builder
A new function replaceToken is created in the tokens' module and integrated into the handleToken function in the unocss-builder class. This new function uses regex to remove specific patterns from the token string, thus providing a way to process and modify the tokens before they are used in the builder.