Add Builder class to generate CSS classes from node attributes
The new class, Builder, is introduced to generate CSS classes from node attributes. The Builder class utilizes the utility function rgbToHex to convert the RGB color value to HEX format. In turn, this is tested in the newly added builder.test.ts file. The Builder class replaces the previous UnocssBuilder. This enables improved definition of attributes and included new type definitions in the types.ts file, specifically for RectSidesNew and RectCornersNew.