Add support for variable utility values in builder
The builder script has been updated to retrieve and handle variable utility values if the config mode is set to 'variables'. This change mainly affects the handling of fill colors, switching to variable utility values when available. A new private method, `getTokenByType`, has been added to facilitate the retrieval of token values of a given type from a scene node.