Refactor nearest color inference to generic nearest inference
The code has been refactored to make the inference of nearest color a more generic "nearest" inference. This change consists of renaming several functions and variables throughout the code to reflect this shift, as well as updating the event messages associated with this function. This broader approach allows for inference in dimensions as well as colors.