Update divider color code in UI component
Changed the divider color code in ui.vue to be determined by a variable. This variable is assigned a different value depending on whether the user is in dark mode or light mode. The new setup improves the readability of the code and enables easier style changes in the future.