chore: add MIT license
This commit introduces an MIT license to the project. The license grants various permissions to users of the software, such as the rights to use, copy, modify, merge, publish, distribute, sublicense, and sell copies of the software. This provides clarity on what users are allowed to do with the project's code.