Add README (#22)
* docs: update README.md with new project description and features

The README.md file is updated to provide a more informative and attractive project description, outlining the features of the "Onyx" Figma plugin. Information about licensing and the project's author have been added.

* chore: decrease logo height in README

* docs: add development section to README.md

This commit includes updates to the README.md file, providing more detailed instructions for the development environment setup. Instructions for installing and verifying Node Version Manager (NVM), Node, and PNPM are now specified, as well as the steps to prepare PNPM cache and .env files.